Move AudioHighlight updating and scrolling to page
